# docs/openapi/frontend-surface.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FERN Frontend Surface",
    "version": "2026-03-31",
    "description": "Source-derived OpenAPI contract for the current frontend-facing gateway surface. It includes endpoints actively used by the frontend app or intentionally exposed for frontend diagnostics. It excludes internal control-plane routes, but now includes the auth session lifecycle, shared paged staff list envelopes with total-count metadata, the stock-safe sales, payroll, and procurement actions already exercised by the frontend, the dedicated cashier POS workspace at exact /pos built on existing protected sales and product endpoints, the protected non-cashier staff order queue at exact /order, and the public customer ordering plus public order-status flow under /api/v1/sales/public."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8180",
      "description": "Local gateway"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/product/products": {
      "get": {
        "summary": "List products",
        "responses": {
          "200": {
            "description": "Product catalog",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductView"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/product/items": {
      "get": {
        "summary": "List items",
        "responses": {
          "200": {
            "description": "Inventory item list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ItemView"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/product/prices": {
      "get": {
        "summary": "List outlet prices effective on a date",
        "parameters": [
          {
            "name": "outletId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "on",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Price list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PriceView"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/api/v1/procurement/suppliers/{supplierId}": {
      "get": {
        "summary": "Get supplier detail",
        "parameters": [
          {
            "name": "supplierId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Supplier detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierView"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/api/v1/procurement/payments": {
      "post": {
        "summary": "Create a supplier payment",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSupplierPaymentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created supplier-payment detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierPaymentView"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "get": {
        "summary": "List supplier payments",
        "parameters": [
          {
            "name": "outletId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "supplierId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startTime",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endTime",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "maximum": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Supplier-payment list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierPaymentListPage"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/v1/procurement/payments/{paymentId}": {
      "get": {
        "summary": "Get supplier-payment detail",
        "parameters": [
          {
            "name": "paymentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Supplier-payment detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierPaymentView"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/v1/procurement/payments/{paymentId}/post": {
      "post": {
        "summary": "Post a supplier payment",
        "parameters": [
          {
            "name": "paymentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Posted supplier-payment detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierPaymentView"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/v1/procurement/payments/{paymentId}/cancel": {
      "post": {
        "summary": "Cancel a supplier payment",
        "parameters": [
          {
            "name": "paymentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cancelled supplier-payment detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierPaymentView"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/v1/procurement/payments/{paymentId}/reverse": {
      "post": {
        "summary": "Reverse a supplier payment",
        "parameters": [
          {
            "name": "paymentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reversed supplier-payment detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierPaymentView"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Request validation or malformed body error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorPayload"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid bearer token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorPayload"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Authenticated but not allowed to access the requested scope",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorPayload"
            }
          }
        }
      },
      "NotFound": {
        "description": "Requested resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorPayload"
            }
          }
        }
      },
...
    },
    "schemas": {
      "ErrorPayload": {
        "type": "object",
        "required": [
          "timestamp",
          "error",
          "message"
        ],
        "properties": {
          "timestamp": {
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {}
        }
      },
...
      "PaymentAllocationRequest": {
        "type": "object",
        "required": [
          "invoiceId",
          "allocatedAmount"
        ],
        "properties": {
          "invoiceId": {
            "type": "integer",
            "format": "int64"
          },
          "allocatedAmount": {
            "type": "number"
          },
          "note": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "CreateSupplierPaymentRequest": {
        "type": "object",
        "required": [
          "supplierId",
          "currencyCode",
          "paymentMethod",
          "amount",
          "paymentTime",
          "allocations"
        ],
        "properties": {
          "supplierId": {
            "type": "integer",
            "format": "int64"
          },
          "currencyCode": {
            "type": "string"
          },
          "paymentMethod": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "paymentTime": {
            "type": "string",
            "format": "date-time"
          },
          "transactionRef": {
            "type": [
              "string",
              "null"
            ]
          },
          "note": {
            "type": [
              "string",
              "null"
            ]
          },
          "allocations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentAllocationRequest"
            }
          }
        }
      },
...
      "ProductView": {
        "type": "object",
        "required": [
          "id",
          "code",
          "name",
          "status"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "categoryCode": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string"
          },
          "imageUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "ItemView": {
        "type": "object",
        "required": [
          "id",
          "code",
          "name",
          "baseUomCode",
          "status"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "categoryCode": {
            "type": [
              "string",
              "null"
            ]
          },
          "baseUomCode": {
            "type": "string"
          },
          "minStockLevel": {
            "type": [
              "number",
              "null"
            ]
          },
          "maxStockLevel": {
            "type": [
              "number",
              "null"
            ]
          },
          "status": {
            "type": "string"
          }
        }
      },
      "PriceView": {
        "type": "object",
        "required": [
          "productId",
          "outletId",
          "currencyCode",
          "priceValue",
          "effectiveFrom"
        ],
        "properties": {
          "productId": {
            "type": "integer",
            "format": "int64"
          },
          "outletId": {
            "type": "integer",
            "format": "int64"
          },
          "currencyCode": {
            "type": "string"
          },
          "priceValue": {
            "type": "number"
          },
          "effectiveFrom": {
            "type": "string"
          },
          "effectiveTo": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "SupplierView": {
        "type": "object",
        "required": [
          "id",
          "supplierCode",
          "name",
          "status",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "regionId": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "supplierCode": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "taxCode": {
            "type": [
              "string",
              "null"
            ]
          },
          "address": {
            "type": [
              "string",
              "null"
            ]
          },
          "phone": {
            "type": [
              "string",
              "null"
            ]
          },
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "contactPerson": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string"
          },
          "deletedAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        }
      },
...
      "SupplierPaymentAllocationView": {
        "type": "object",
        "required": [
          "invoiceId",
          "allocatedAmount"
        ],
        "properties": {
          "invoiceId": {
            "type": "integer",
            "format": "int64"
          },
          "allocatedAmount": {
            "type": "number"
          },
          "note": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "SupplierPaymentView": {
        "type": "object",
        "required": [
          "id",
          "supplierId",
          "currencyCode",
          "paymentMethod",
          "amount",
          "status",
          "paymentTime",
          "createdAt",
          "updatedAt",
          "allocations"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "supplierId": {
            "type": "integer",
            "format": "int64"
          },
          "currencyCode": {
            "type": "string"
          },
          "paymentMethod": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "status": {
            "type": "string"
          },
          "paymentTime": {
            "type": "string"
          },
          "transactionRef": {
            "type": [
              "string",
              "null"
            ]
          },
          "note": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdByUserId": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "allocations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SupplierPaymentAllocationView"
            }
          }
        }
      },
...
      "SupplierPaymentListItemView": {
        "type": "object",
        "required": [
          "id",
          "supplierId",
          "outletId",
          "currencyCode",
          "paymentMethod",
          "amount",
          "status",
          "paymentTime",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "supplierId": {
            "type": "integer",
            "format": "int64"
          },
          "outletId": {
            "type": "integer",
            "format": "int64"
          },
          "currencyCode": {
            "type": "string"
          },
          "paymentMethod": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "status": {
            "type": "string"
          },
          "paymentTime": {
            "type": "string"
          },
          "transactionRef": {
            "type": [
              "string",
              "null"
            ]
          },
          "note": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdByUserId": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        }
      },
...
      "PagedResultMeta": {
        "type": "object",
        "required": [
          "limit",
          "offset",
          "totalCount",
          "hasNextPage"
        ],
        "properties": {
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          },
          "totalCount": {
            "type": "integer",
            "format": "int64"
          },
          "hasNextPage": {
            "type": "boolean"
          }
        }
      },
...
      "SupplierPaymentListPage": {
        "allOf": [
          {
            "$ref": "#/components/schemas/PagedResultMeta"
          },
          {
            "type": "object",
            "required": [
              "items"
            ],
            "properties": {
              "items": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/SupplierPaymentListItemView"
                }
              }
            }
          }
        ]
      },
...
    }
  }
}